backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.117.1"}
psycopg = {extras = ["binary"], version = "^3.2.10"}
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
redis = "^6.4.0"
websockets = "^15.0.1"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.1"
pydantic-settings = "^2.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
